// kaphor-frontend/src/constants/market.d.ts
export declare const MARKET_CATEGORIES: {
    group: string;
    items: string[];
}[];
export declare const ALL_CATEGORY_ITEMS: string[];
export declare const ACCESSORY_CATEGORY_ITEMS: string[];
export declare function isAccessoryCategory(category?: string | null, subCategory?: string | null): boolean;
export declare const MARKET_CONDITIONS: {
    id: string;
    label: string;
    desc: string;
}[];
export declare const LISTING_TYPES: {
    id: string;
    label: string;
    desc: string;
}[];
export declare const MARKET_SIZES: string[];
//# sourceMappingURL=market.d.ts.map